Getting the client side code to compile
Most of the client side code was originally written in separate
segments, without actually checking that it compiled. So a bunch of the
function calls had the wrong names, or the linker freaked out because we
weren't cross compiling to C++ correctly, a bunch of stuff like that.
This commit fixes those "can't compile" issues.

Note that this commit doesn't guarantee that the code _works_ at all,
just that it now compiles properly. Still waiting on a test bed to
ensure that it actually works (either a test script or another aruduino
to test on).

There's also some functionality that isn't in here, this only includes
the core required FSM and button polling logic. LCD driver, seven
segment driver, and SD card driver all still need writing.