Add 3 battery holders to tower side backplane
Assuming that we're running the same batteries (they're cheap as hell
and worked awesome in IREC 2017). The system could easily be run off of
a single battery (they hold a lot of charge), but by having multiple we
can run the actuators off of a separate power supply than the arduino
and xbee. Three should be more than overkill, one holder might not even
need to be populated.