Optimize findDigits function, change a couple of types
Comments for Aidan:
The seven segment display can only display numbers from 0x00 to 0xFF,
which fortunately is exactly the range of a uint8_t. By switching from
long to uint8_t, we remove the need to check that numToShow is within
the range of 0-255, since that's guaranteed by the use of 8 bits.

Sinc digits is a member of the SevSeg struct, you don't really need to
pass both a digits array and a SevSeg pointer to setNewNum, you can away
with just passing the SevSeg pointer and taking digits[] out of that.

Since we're hardcoding 2 digits into SevSeg, you don't need the for loop
in findDigits, you can just assume that it's 2.

Division is really expensive, but since we're dividing by 16 you can
just logical right shift by 4 (they might have mentioned this in ECE
150), which should run faster. And to take the bottom 4 bits, we can
just logical and with 0x0F. I probably can't explain bitwise operations
in a git commit message, but let me know if you need any elaboration.