Remove a few more lines based on hardcoding 2 digits
Comments for Aidan:
I removed numDigits from the sevSegDisplay struct since it's now assumed
to be 2. I also slightly rearranged the for loop in refresh, so it no
longer needs to check the value of i to determine which digit to turn
on.

I think those are all of the changes. I also removed the numDigitsIn
argument from start() function, since it's assumed to be 2.