Add test cut to check sizes and footprints
Needed a smaller thing I could cut to ensure that the footprints were
all the right size, the holes were proper size, the grommets fit, and
the zip tie thing held wires.

After making this test cut on 2018-01-19, can confirm that the arduino
mega footprint, the power board footprint, the battery holder footprint
is correct, and the zip tie cutouts work a treat. Issue, however, is
that the holes (1.5mm radius) are too big (actually, they make
excellent clearance holes, but terrible tapped holes), so for all tapped
holes we should take them down to 1.25mm radius.

Test cut looks good, proceeding with real cut so that I can assemble the
box and bring it to _all_ the interviews.