Add daq polling to tower main loop
Dummy values for calibration at the moment. Need to map a 10bit ADC
value with 5V full scale into a mass, 2 pressures, and current stuff.
Calibration to be done eventually. Daq values are not logged at present.

tested with limit switches and valves, and that part seems to work
great, still need to actually hook up sensors and see how that handles.