Start work on nodeio.ioio
This is largely an API for the flight_instr people. It provides an
interface for sending data back to the tower RLCS box, which then sends
things back to the operator. The FSM and protocol stuff is largely based
on RLSC's stuff.

This commit is not functional, but it's the first time anything worked.
Currently the vent valve communication works. Still need to write the
injector valve control code, and the sensor packing code, and all of the
tower side control code.