Get node_inj's actuator working
Copy pasted the FSM code from the vent valve over to the injector valve.
It works, so yay! Still need to do a bunch of tower stuff.

This commit contains some tower handling code, eg the FSM part for
recieving sensor data strings (still no unpacking of those). That code
is in this commit. Should it be in its own commit? Yes. Yes it should.