Finish unpack sensor data, unit test now passes
The sensor decoding is kinda wonky (by design). It only allows 9 bits of
precision from the pressure reading (bits 9:1 of that parameter) and 6
bits of precision from the thermistor readings (bits 9:3). Make sure to
scale your values appropriately. This is fixeable, it's just that way to
reduce number of bytes that have to go over the radio link. If we decide
that the full 10 bits of ADC precision are absolutely necessary, that
can be added in.