Change tower to use Serial1 for XBEE
I got tired of being unable to reprogram the arduino while the xbee was
plugged in, so I moved it to another Serial interface. This also fixes
the issue where on startup, the tower arduino would see bytes available
on its serial lines, and think it was being reprogrammed, and not leave
the bootloader. So that's nice