S19WDR1 slightly below minimum viable product
RLCS now kinda interoperates with RocketCAN. Nodeio.ioio mostly works, but
isn't deserializing errors and there are probably bugs I haven't seen. The
changes to nodeio.ioio include bug fixes and some minor code addition.

Forced the tower side box to tell rocketCAN to depower the system bus if
ever the client side box stops talking (safe state protocol now has that
bus depowered). I think this roughly maps to what we want. If an operator
is in control of the rocket, then everything in the rocket should be alive.
To make this happen, the system_state type needed redefinition to include a
is_bus_powered field. This replaced running_self_test, which wasn't used.

sd_handler is now all kinds of fucked up. That needs to be fixed, but it
doesn't need to be fixed before WDR.

tower_handle_*_update in tower_globals needs fixing, that was all removed
to deal with not having two discrete boards in the rocket.