Fix bug that opened injector valve on RLCS startup
The tower side box's desired rocketCAN state wasn't being initialized,
so everything was set to 0. That caused both valves to default to OPEN,
which is great for vent, but suboptimal for injector.

This commit forces injector to NOTHING on RLCS startup, which causes the
valve to be depowered, which is all of what you want, and none of what
you don't.