Fix DAQ readings on the pressure sensors
There was an issue with the sensor SP1 in which it would not report any
pressures below 500psi (reporting all of them as 500psi). Debugging this
issue, it was found to be that one of the pins on the arduino (pin A6)
had failed in some way, and was providing a voltage of 2.5V (likely from
a diode to VDD). This problem was fixed by changing the SP1 pin to A4.

This commit also adjusts the pressure sensor calibration constants, as
found on a somewhat sketchy 5 point manual pressure guage calibration.
Should be good for WDR, we may want to recalibrate for IREC.